requirements.txt
numpy
scipy
matplotlib
torch==0.4.1 # thay thế bằng phiên bản mới nhất
torchvision==0.2.1 # thay thế bằng phiên bản mới nhất
nltk
Pillow
tqdm
lmdb
tensorflow
tensorboardX
streamlit
